Fix HeaderWave component by separating VertexDotsObject
